[0.4.2] - Tag the release

+++ reactos/sdk/lib/crt/float/i386/statfp.c
@@ -25,7 +25,7 @@
 unsigned int CDECL _statusfp(void)
 {
   unsigned int retVal = 0;
-  unsigned short fpword;
+  unsigned int fpword;
 
 #ifdef _M_AMD64
     fpword = _mm_getcsr();
